# app/internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.CafeTableResponse:
    properties:
      capacity:
        example: 4
        type: integer
      created_at:
        type: string
      hall_id:
        example: a1b2c3d4-e5f6-4a5b-8c9d-e0f1a2b3c4d5
        type: string
      id:
        example: c0f18a64-7f5c-4425-9414-1b01cddee9d9
        type: string
      number:
        example: 5
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/model.TableStatus'
        example: free
      updated_at:
        type: string
    type: object
...
  model.CreateCafeTableRequest:
    properties:
      capacity:
        example: 4
        minimum: 1
        type: integer
      hall_id:
        example: a1b2c3d4-e5f6-4a5b-8c9d-e0f1a2b3c4d5
        type: string
      number:
        example: 5
        minimum: 1
        type: integer
      status:
        example: free
        type: string
    required:
    - capacity
    - hall_id
    - number
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  model.UpdateCafeTableRequest:
    properties:
      capacity:
        example: 4
        type: integer
      hall_id:
        example: a1b2c3d4-e5f6-4a5b-8c9d-e0f1a2b3c4d5
        type: string
      number:
        example: 5
        type: integer
      status:
        example: available
        type: string
    type: object
...
host: back.staging.maryai.yurtal.tech
info:
  contact: {}
  description: MaryAI API server with multi-language support (uz, ru, en)
  title: MaryAI API
  version: "1.0"
paths:
...
  /api/v1/cafe-tables:
    get:
      consumes:
      - application/json
      description: Get all cafe tables with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all cafe tables
      tags:
      - cafe-tables
    post:
      consumes:
      - application/json
      description: Create a new cafe table in a specific hall
      parameters:
      - description: Create Cafe Table Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateCafeTableRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new cafe table
      tags:
      - cafe-tables
  /api/v1/cafe-tables/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a cafe table
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Bearer: []
      - BearerAuth: []
      summary: Delete a cafe table
      tags:
      - cafe-tables
    get:
      consumes:
      - application/json
      description: Get a specific cafe table by its ID
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get cafe table by ID
      tags:
      - cafe-tables
    put:
      consumes:
      - application/json
      description: Update a cafe table details
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Cafe Table Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCafeTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Bearer: []
      - BearerAuth: []
      summary: Update a cafe table
      tags:
      - cafe-tables
  /api/v1/cafe-tables/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft deleted cafe table
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a cafe table
      tags:
      - cafe-tables
  /api/v1/cafe-tables/{id}/set-busy:
    post:
      consumes:
      - application/json
      description: Mark a cafe table as busy (occupied)
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set table as busy
      tags:
      - cafe-tables
  /api/v1/cafe-tables/{id}/set-free:
    post:
      consumes:
      - application/json
      description: Mark a cafe table as free (available)
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set table as free
      tags:
      - cafe-tables
  /api/v1/cafe-tables/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a cafe table
      parameters:
      - description: Cafe Table ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Cafe Table Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCafeTableStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CafeTableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - Bearer: []
      - BearerAuth: []
      summary: Update cafe table status
      tags:
      - cafe-tables
  /api/v1/cafe-tables/available/capacity:
    get:
      consumes:
      - application/json
      description: Get available tables that can accommodate a minimum number of guests
      parameters:
      - description: Minimum capacity
        in: query
        name: capacity
        required: true
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available tables by capacity
      tags:
      - cafe-tables
  /api/v1/cafe-tables/available/hall/{hall_id}:
    get:
      consumes:
      - application/json
      description: Get all available (free) tables in a specific hall
      parameters:
      - description: Hall ID
        in: path
        name: hall_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available tables by hall
      tags:
      - cafe-tables
  /api/v1/cafe-tables/available/hall/{hall_id}/capacity:
    get:
      consumes:
      - application/json
      description: Get available tables in a specific hall that can accommodate a
        minimum number of guests
      parameters:
      - description: Hall ID
        in: path
        name: hall_id
        required: true
        type: string
      - description: Minimum capacity
        in: query
        name: capacity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available tables by hall and capacity
      tags:
      - cafe-tables
  /api/v1/cafe-tables/hall-status:
    get:
      consumes:
      - application/json
      description: Get tables in a specific hall with a specific status
      parameters:
      - description: Hall ID
        in: query
        name: hall_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get cafe tables by hall and status
      tags:
      - cafe-tables
  /api/v1/cafe-tables/hall/{hall_id}:
    get:
      consumes:
      - application/json
      description: Get all tables in a specific hall
      parameters:
      - description: Hall ID
        in: path
        name: hall_id
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CafeTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get cafe tables by hall ID
      tags:
      - cafe-tables
...
  /api/v1/cafe-tables/stats/occupancy:
    get:
      consumes:
      - application/json
      description: Get overall cafe table occupancy statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TableOccupancyStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get table occupancy statistics
      tags:
      - cafe-tables
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"